+++ src/views/loans/myHistory/Index.vue
@@ -288,7 +288,7 @@
                         "id": 2,
                         "name": "myLoans.active",
                         "onDisplay": "Aktif",
-                        "icon": "fa fa-clock-o"
+                        "icon": "fa fa-hourglass-start"
                     },
                     {
                         "id": 3,
@@ -451,6 +451,42 @@
                     this.buttonDisabled = false;
                 }
                 this.isLoading = false;
+            },
+            detailRouter(id){
+                // console.log("Teset")
+                this.setProgress = true;
+                this.isLoadingRouter = true;
+                this.secondaryButtonDisabled = true;
+                this.submitEnabled = false;
+                this.buttonDisabled = true;
+                let data = window.btoa(id);
+                try{
+                    if(this.setProgress == true) {
+                        this.intervalProgressbar = setInterval(() => {
+                            this.widthProgressBar += 35;
+                            this.widhtStyle = "width: "+ this.widthProgressBar.toString() +"%;";
+                            // console.log(this.widhtStyle);
+                        }, 1000);
+                        if(this.widthProgressBar == 100) {
+                            clearInterval(this.intervalProgressbar);
+                            this.widthProgressBar = 0;
+                            this.setProgress == false;
+                            // this.setProgress = false;
+                        }
+                        // console.log("Test");
+                        setTimeout(() => {
+                            this.$router.push({ name: 'myLoans.details', query: {data: data} }).then(() => { this.$router.go() })
+                        }, 4000);
+                    }
+                } catch(e) {
+                    this.errorResponse = [
+                        {
+                            'id': 1,
+                            'message': 'Error!', 
+                            'detail': e,
+                        }
+                    ];
+                }
             },
             createLoans(){
                 this.setProgress = true;
