+++ src/views/loans/myHistory/ReturnDetails.vue
@@ -85,7 +85,7 @@
                         <div class="container-fluid">
                             <!-- DataTales Example -->
                             <div :class= "windowWidth <= $widthPotraitPhone ? 'container my-5 p-5' : 'container my-5 p-5 shadow-lg bg-body rounded'">
-                                <div :class="this.loansStatus == '3' ? 'row':'d-none'">
+                                <div v-if="this.detailObject.loaner_id == this.$session.id" class="row">
                                     <div v-if="isLoadingImage == true" class="col-sm-12 d-md-none text-center my-5">
                                         <div v-if="windowWidth < 720">
                                             <div class="m-3 spinner-grow spinner-grow-sm text-secondary" role="status">
@@ -323,7 +323,7 @@
                                         </div>
                                     </div>
                                 </div>
-                                <div :class="this.loansStatus == '3' ? 'd-none':'row mh-100'">
+                                <div :class="this.detailObject.loaner_id == this.$session.id ? 'd-none':'row mh-100'">
                                     <div v-if="this.windowWidth >= this.$widthLandscapePhone" class="col-3">&nbsp;
                                     </div>
                                     <div v-if="this.windowWidth >= this.$widthLandscapePhone" class="col-4 mx-5">
@@ -335,8 +335,8 @@
                                     <div v-else class="col-12">
                                         <img  class="w-100 img-thumbnails" :src="this.$baseUrl+'/src/assets/img/404.png'" alt="">
                                     </div>
-                                    <h3 class="text-center my-3">Status tidak sesuai</h3>
-                                    <p class="text-center my-3">Status data peminjaman bukan permintaan peminjaman. Permintaan perbaruan data tidak dapat dilakukan</p>
+                                    <h3 class="text-center my-3">Data Tidak Ditemukan</h3>
+                                    <p class="text-center my-3">Permintaan tidak dapat dilakukan</p>
                                 </div>
                                 <div class="row my-4 d-flex justify-content-center">
                                     <div :class="this.windowWidth <= $widthLandscapePhone ? 'col-12' :'col-4'">
@@ -571,7 +571,7 @@
                         }
                         // console.log("Test");
                         setTimeout(() => {
-                            this.$router.push({ name: 'manageLoans.done' }).then(() => { this.$router.go() })
+                            this.$router.push({ name: 'myLoans.done' }).then(() => { this.$router.go() })
                         }, 4000);
                     }
                 } catch(e) {
@@ -609,7 +609,7 @@
                         }
                         // console.log("Test");
                         setTimeout(() => {
-                            this.$router.push({ name: 'manageLoans.details', query: {data: data} }).then(() => { this.$router.go() })
+                            this.$router.push({ name: 'myLoans.details', query: {data: data} }).then(() => { this.$router.go() })
                         }, 4000);
                     }
                 } catch(e) {
